Move countdown clock into its own component
